updated changes

+++ app/Http/Controllers/settings/RolesController.php
@@ -41,26 +41,33 @@ class RolesController extends Controller
      * @return \Illuminate\Http\Response
      */
 
-     public function roleNameValidation(Request $request)
-     {
-       
-         $model = Role::where('name',$request->name)->where('id', '!=', $request->id)->get();
-//   dd($model);
-         $response = (count($model))?false:true;
-         return response()->json(['response'=>$response]);
-     }
+    public function roleNameValidation(Request $request)
+    {
+
+        $model = Role::where('name', $request->name)->where('id', '!=', $request->id)->get();
+        //   dd($model);
+        $response = (count($model)) ? false : true;
+        return response()->json(['response' => $response]);
+    }
     public function store(Request $request)
     {
 
         Log::info('RoleController->Store:-Inside ' . json_encode($request->all()));
         $input = $request->all();
+        $ispermission = (isset($request->permission)) ? true : false;
+
+        if (!$ispermission) {
+            $input['permission'] = ['dashboard-view'];
+        }
+
         $input['guard_name'] = 'web';
         $permissionitems = array();
         $permissionsDatas = $input['permission'];
-       
+
+
         Log::info('ProjectController->Store:-Inside $permissionsDatas' . json_encode($permissionsDatas));
         foreach ($permissionsDatas as $permissionsData) {
-           
+
             Log::info('ProjectController->Store:-Inside $permissionsData' . json_encode($permissionsData));
             $permissionitems[] = Permission::where('name', $permissionsData)->first()->id;
             Log::info('ProjectController->Store:-Inside permissionitems[]' . json_encode($permissionitems));
@@ -73,42 +80,38 @@ class RolesController extends Controller
 
 
         $permissionsNew = explode(",", $permissionitems1);
-      
+
 
         unset($input['_token']);
         $id = isset($request->id) ? $request->id : '';
-       
+
         if ($id) {
             $role = Role::find($id);
-        } else {         
+        } else {
 
             $role = new Role();
         }
-       
-       $role->name =$input['name'];
-       $role->guard_name =$input['guard_name'];
-       $role->authority_type =$input['authority_type'];
-       $role->save();
-      
-        if (isset($permissionsDatas)) {
-         
-               $role->syncPermissions($permissionsNew); 
-              
-             
-            } else {
 
-               //If no role is selected remove exisiting permissions associated to a role
-               $p_all = Permission::all(); //Get all permissions
-               foreach ($p_all as $p) {
-                  $role->revokePermissionTo($p); //Remove all permissions associated with role
-               }
+        $role->name = $input['name'];
+        $role->guard_name = $input['guard_name'];
+        $role->authority_type = $input['authority_type'];
+        $role->save();
+
+        if (isset($permissionsDatas)) {
+
+            $role->syncPermissions($permissionsNew);
+        } else {
+
+            //If no role is selected remove exisiting permissions associated to a role
+            $p_all = Permission::all(); //Get all permissions
+            foreach ($p_all as $p) {
+                $role->revokePermissionTo($p); //Remove all permissions associated with role
             }
+        }
         if ($role) {
             return redirect('roles')->with('success', "Roles Stored successfully.");
         } else {
             return redirect()->back()->withErrors(['error' => ['Insert Error']]);
-
-           
         }
     }
 
@@ -131,8 +134,8 @@ class RolesController extends Controller
      */
     public function edit($id)
     {
-        $roles=Role::where('id',$id)->first();
-        return view('Settings.Role.edit',['roles'=>$roles]);
+        $roles = Role::where('id', $id)->first();
+        return view('Settings.Role.edit', ['roles' => $roles]);
     }
 
     /**
@@ -155,8 +158,24 @@ class RolesController extends Controller
      */
     public function destroy($id)
     {
-        $department_update = Role::where("id", $id)->delete();
-        echo json_encode($department_update);
+
+        $role = Role::find($id);
+
+        if ($role->hasPermissionTo($role->id)) {
+            return response()->json([
+                "status" => "failed",
+                'message' => 'You cannot delete a role that has permissions assigned!'
+            ]);
+            
+        } else {
+            $department_update = Role::where("id", $id)->delete();
+            return response()->json([
+                "status" => "success",
+                'message' => 'Deleted Success!'
+            ]);
+        }
+        // $department_update = Role::where("id", $id)->delete();
+        // echo json_encode($department_update);
     }
     public function search(Request $request)
     {
